Update Source.cpp
Oops, I goofed up last time. Not only does it click as fast as it can (with grinder mode), it is also unstoppable. The only opportunity to stop it is when it sleeps 2 seconds to eat. However, that would mean your right mouse key was always held down. Sorry for the inconvenience, better test it before I put it on here!